lovelove/requirement.txt
amqp==2.5.0
anyjson==0.3.3
billiard==3.6.0.0
celery==4.3.0
celery-with-redis==3.0
Click==7.0
Django==1.11.16
django-celery-results==1.1.2
django-simple-captcha==0.4.6
django-tinymce==2.8.0
djangorestframework==3.9.4
Flask==1.1.0
itsdangerous==1.1.0
Jinja2==2.10.1
kombu==4.6.3
MarkupSafe==1.1.1
mysqlclient==1.4.2.post1
pbr==5.3.1
Pillow==6.1.0
PyMySQL==0.9.3
pytz==2019.1
redis==3.2.1
six==1.12.0
stevedore==1.30.1
vine==1.3.0
virtualenv==16.6.1
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
virtualenvwrapper-win==1.2.5
Werkzeug==0.15.4
